examples/workflow-demo/gpt-workflow-agent/requirements.txt
openai==1.58.1
python-dotenv==1.0.0
redis==5.0.1